requirements.txt
qrcodes
FPDF
